rocm: add unique metric utility functions for intercept mode
In intercept mode we are only interested in unique events, i.e., events
that have the same name and instance (can be from different devices).
This is because in intercept mode all unique events are monitored on all
devices. Though, only the counters for the actual requested events will
be presented to the user. This is a design choice that accounts for the
fact that once set, callbacks for dispatch queues cannot be updated
(this includes the monitored events).