rocm: add qualifier support
This commit contains the core changes of this feature set. It introduces
the logic necessary to handle event identifiers in such a way that
device and instance attributes are presented to the PAPI users as
qualifiers. This means that papi_native_avail will return:

Native Events in Component: rocm
===============================================================================
| rocm:::SQ_WAIT_INST_LDS                                                      |
|            Number of wave-cycles spent waiting for LDS instruction issue. In |
|            units of 4 cycles. (per-simd, nondeterministic)                   |
|     :device=0                                                                |
|             mandatory device qualifier [devices: 0,1]                        |
--------------------------------------------------------------------------------
| rocm:::TCP_TCP_TA_DATA_STALL_CYCLES                                          |
|            TCP stalls TA data interface. Now Windowed.                       |
|     :device=0                                                                |
|             mandatory device qualifier [devices: 0,1]                        |
|     :instance=0                                                              |
|             mandatory instance qualifier in range [0 - 15]                   |
--------------------------------------------------------------------------------

The PAPI user will be able to use event names in the same form as
before (all previous tests will still work) with the relaxation on the
order of device and instance numbers.